+++ packages/backend/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './logger';
-export * from './ajv';
 export * from './normalizeUrl';
 export * from './env';
+export * from './typebox';