+++ packages/frontend/components/collection/add/index.ts
@@ -1,1 +1,2 @@
 export * from './AddCollectionForm';
+export * from './AddCollectionModal';