+++ packages/frontend/components/sidebar/Chevron.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 import { BsChevronRight } from '@react-icons/all-files/bs/BsChevronRight';