+++ packages/frontend/features/AddCollectionModal/index.ts
@@ -1,2 +1,0 @@
-export * from './AddCollectionModal';
-export * from './useAddCollectionModal';