+++ packages/frontend/features/Panes/AddModal.tsx
@@ -1,30 +1,36 @@
-import { useCallback, useState, useEffect } from 'react';
-import { useDisclosure, useToast } from '@chakra-ui/react';
+import { useCallback, useEffect } from 'react';
+import { useContextSelector } from 'use-context-selector';
+import { useToast } from '@chakra-ui/react';
+import {
+  AddCollectionFormData,
+  AddCollectionModal,
+} from '@components/collection/add';
 import { useCollectionsTree } from '@features/Collections';
+import { defaultIcon, emptyIfUndefined, ID } from '@orpington-news/shared';
+import { ModalContext } from './ModalContext';
 import {
-  Collection,
-  defaultIcon,
-  emptyIfUndefined,
-  ID,
-} from '@orpington-news/shared';
-import { AddCollectionFormData } from '@components/collection/add';
-import {
-  useEditCollection,
-  useSaveCollection,
   useVerifyFeedURL,
+  useSaveCollection,
+  useEditCollection,
 } from './queries';
 
-interface State {
+export interface AddModalState {
   isUrlVerified: boolean;
   initialData?: AddCollectionFormData;
   editedFeedId?: ID;
 }
 
-export const useAddCollectionModal = () => {
+export const AddModal: React.FC = () => {
   const toast = useToast();
-  const [state, setState] = useState<State>({ isUrlVerified: false });
 
-  const { isOpen, onClose, onOpen } = useDisclosure();
+  const state = useContextSelector(ModalContext, (ctx) => ctx.addModalState);
+  const setState = useContextSelector(
+    ModalContext,
+    (ctx) => ctx.setAddModalState
+  );
+  const isOpen = useContextSelector(ModalContext, (ctx) => ctx.isAddModalOpen);
+  const onClose = useContextSelector(ModalContext, (ctx) => ctx.closeAddModal);
+
   const { mutate: verifyFeedURL, isLoading: isVerifying } = useVerifyFeedURL();
   const { mutate: saveCollection, isLoading: isSaving } = useSaveCollection({
     onSuccess: () => {
@@ -63,7 +69,7 @@ export const useAddCollectionModal = () => {
         };
       });
     },
-    [state.initialData?.url]
+    [setState, state.initialData?.url]
   );
 
   const onVerifyUrlClicked = useCallback(
@@ -85,23 +91,7 @@ export const useAddCollectionModal = () => {
         }
       );
     },
-    [verifyFeedURL]
-  );
-
-  const onOpenAddCollectionModal = useCallback(
-    (initialData?: Collection) => {
-      if (initialData) {
-        setState({
-          isUrlVerified: true,
-          initialData,
-          editedFeedId: initialData!.id,
-        });
-      } else {
-        setState({ isUrlVerified: false });
-      }
-      onOpen();
-    },
-    [onOpen]
+    [setState, verifyFeedURL]
   );
 
   useEffect(() => {
@@ -110,21 +100,21 @@ export const useAddCollectionModal = () => {
         isUrlVerified: false,
       });
     }
-  }, [isOpen]);
+  }, [isOpen, setState]);
 
-  return {
-    onOpenAddCollectionModal,
-    isOpen,
-    onClose,
-
-    modalTitle: state.editedFeedId ? 'Edit feed' : 'Add feed',
-    initialData: state.initialData,
-    isUrlVerified: state.isUrlVerified,
-    onVerifyUrlChanged,
-    onVerifyUrlClicked,
-    areCollectionsLoading,
-    collections: emptyIfUndefined(collections),
-    isLoading: isVerifying || isSaving,
-    onSubmit: state.editedFeedId ? editCollection : saveCollection,
-  };
+  return (
+    <AddCollectionModal
+      isOpen={isOpen}
+      onClose={onClose}
+      modalTitle={state.editedFeedId ? 'Edit feed' : 'Add feed'}
+      initialData={state.initialData}
+      isUrlVerified={state.isUrlVerified}
+      onVerifyUrlChanged={onVerifyUrlChanged}
+      onVerifyUrlClicked={onVerifyUrlClicked}
+      areCollectionsLoading={areCollectionsLoading}
+      collections={emptyIfUndefined(collections)}
+      isLoading={isVerifying || isSaving}
+      onSubmit={state.editedFeedId ? editCollection : saveCollection}
+    />
+  );
 };