+++ packages/frontend/features/Panes/Panes.tsx
@@ -1,68 +1,32 @@
 import { useCallback } from 'react';
 import { useRouter } from 'next/router';
-import { useLocalStorage } from 'usehooks-ts';
 import { Panes as PanesComponent } from '@components/panes';
-import { CollectionLayout, ID } from '@orpington-news/shared';
 import { Article } from '@features/Article';
-import {
-  useRefreshCollection,
-  useCollectionsContext,
-  useSetCollectionLayout,
-} from '@features/Collections';
-import {
-  DeleteCollectionModal,
-  useDeleteCollectionModal,
-} from '@features/DeleteCollectionModal';
-import { useActiveCollection } from '@features/Preferences';
-import { getNumber } from '@utils/router';
-import { ReactFCC } from '@utils/react';
-import { useDisclosure } from '@chakra-ui/react';
-import { Sidebar } from './Sidebar';
+import { ReactFCC, getNumber, useCookie } from '@utils';
+import { ArticleWidth } from '@orpington-news/shared';
 import { CollectionItemsList } from './CollectionItemsList';
-import { useCookie } from '@utils';
+import { Sidebar } from './Sidebar';
+import { ModalContextProvider } from './ModalContext';
+import { CollectionItemsHeader } from './CollectionItemsHeader';
+import { Drawer } from './Drawer';
+import { AddModal } from './AddModal';
+import { DeleteModal } from './DeleteModal';
 
 interface PanesProps {
   sidebarWidthValue?: number;
   collectionItemsWidthValue?: number;
+  articleWidthValue?: ArticleWidth;
 }
 
 export const Panes: ReactFCC<PanesProps> = ({
   sidebarWidthValue,
   collectionItemsWidthValue,
+  articleWidthValue,
   children,
 }) => {
   const router = useRouter();
   const collectionId = getNumber(router.query?.collectionId);
   const itemId = getNumber(router.query?.itemId);
-
-  const { onOpenDeleteCollectionModal, ...deleteCollectionModalProps } =
-    useDeleteCollectionModal();
-
-  const { activeCollection } = useActiveCollection();
-  const { currentlyUpdatedCollections } = useCollectionsContext();
-  const { mutate: refreshCollection } = useRefreshCollection();
-
-  const handleRefreshClicked = useCallback(
-    (collectionId: ID | string) => {
-      if (typeof collectionId === 'string' && collectionId !== 'home') {
-        return;
-      }
-
-      refreshCollection({ id: collectionId });
-    },
-    [refreshCollection]
-  );
-
-  const { mutate: setCollectionLayout } = useSetCollectionLayout();
-  const handleCollectionLayoutChanged = useCallback(
-    (layout: CollectionLayout) => {
-      setCollectionLayout({
-        id: activeCollection.id,
-        layout,
-      });
-    },
-    [activeCollection.id, setCollectionLayout]
-  );
 
   const handleGoBack = useCallback(() => {
     router.push('/');
@@ -77,27 +41,12 @@ export const Panes: ReactFCC<PanesProps> = ({
     collectionItemsWidthValue ?? 400
   );
 
-  const {
-    isOpen: isDrawerOpen,
-    onClose: onCloseDrawer,
-    onToggle: onToggleDrawer,
-  } = useDisclosure();
-
   return (
-    <>
+    <ModalContextProvider>
       <PanesComponent
         flexGrow={1}
-        isDrawerOpen={isDrawerOpen}
-        onCloseDrawer={onCloseDrawer}
-        onToggleDrawer={onToggleDrawer}
-        activeCollection={activeCollection}
-        currentlyUpdatedCollections={currentlyUpdatedCollections}
-        sidebar={
-          <Sidebar
-            onCloseDrawer={onCloseDrawer}
-            onOpenDeleteCollectionModal={onOpenDeleteCollectionModal}
-          />
-        }
+        sidebar={<Sidebar />}
+        collectionItemHeader={<CollectionItemsHeader />}
         collectionItemList={<CollectionItemsList />}
         mainContent={
           itemId &&
@@ -106,20 +55,22 @@ export const Panes: ReactFCC<PanesProps> = ({
               collectionId={collectionId}
               itemId={itemId}
               onGoBackClicked={handleGoBack}
+              articleWidthValue={articleWidthValue}
             />
           )
         }
-        onRefreshClicked={handleRefreshClicked}
         sidebarWidth={sidebarWidth}
         onSidebarWidthChanged={setSidebarWidth}
         collectionItemsWidth={collectionItemsWidth}
         onCollectionItemsWidthChanged={setCollectionItemsWidth}
-        onCollectionLayoutChanged={handleCollectionLayoutChanged}
       />
 
-      <DeleteCollectionModal {...deleteCollectionModalProps} />
+      <AddModal />
+      <DeleteModal />
+
+      <Drawer sidebar={<Sidebar />} />
 
       {children}
-    </>
+    </ModalContextProvider>
   );
 };